Changes according to Parik’s feedback
