+++ s3/src/main/resources/reference.conf
@@ -6,4 +6,5 @@ akka.stream.alpakka.s3 {
   disk-buffer-path = ""
 
   debug-logging = false
+  
 }
