--- s3/src/main/scala/akka/stream/alpakka/s3/impl/package.scala
@@ -1,5 +1,0 @@
-package akka.stream.alpakka.s3
-
-package object impl {
-  type UploadID = String
-}
