+++ s3/src/main/scala/akka/stream/alpakka/s3/javadsl/S3Client.scala
@@ -5,7 +5,7 @@ import akka.http.impl.model.JavaUri
 import akka.http.javadsl.model.Uri
 import akka.stream.Materializer
 import akka.stream.alpakka.s3.impl.CompleteMultipartUploadResult
-import akka.stream.alpakka.s3.signing.AWSCredentials
+import akka.stream.alpakka.s3.auth.AWSCredentials
 import akka.stream.alpakka.s3.impl.S3Stream
 import akka.stream.alpakka.s3.impl.S3Location
 import akka.stream.javadsl.Source
@@ -15,13 +15,13 @@ import akka.stream.javadsl.Sink
 import java.util.concurrent.CompletionStage
 import scala.compat.java8.FutureConverters._
 
-case class MultipartUploadResult(location: Uri, bucket: String, key: String, etag: String)
+final case class MultipartUploadResult(location: Uri, bucket: String, key: String, etag: String)
 
 object MultipartUploadResult {
-  def apply(r: CompleteMultipartUploadResult) = new MultipartUploadResult(JavaUri(r.location), r.bucket, r.key, r.etag)
+  def apply(r: CompleteMultipartUploadResult): MultipartUploadResult = new MultipartUploadResult(JavaUri(r.location), r.bucket, r.key, r.etag)
 }
 
-class S3Client(credentials: AWSCredentials, region: String, system: ActorSystem, mat: Materializer) {
+final class S3Client(credentials: AWSCredentials, region: String, system: ActorSystem, mat: Materializer) {
   private val impl = new S3Stream(credentials, region)(system, mat)
 
   def download(bucket: String, key: String): Source[ByteString, NotUsed] = impl.download(S3Location(bucket, key)).asJava
