+++ s3/src/main/scala/akka/stream/alpakka/s3/scaladsl/S3Client.scala
@@ -5,7 +5,7 @@ import akka.actor.ActorSystem
 import akka.stream.Materializer
 import akka.stream.alpakka.s3.impl.S3Location
 import akka.stream.alpakka.s3.impl.S3Stream
-import akka.stream.alpakka.s3.signing.AWSCredentials
+import akka.stream.alpakka.s3.auth.AWSCredentials
 import akka.stream.scaladsl.Source
 import akka.util.ByteString
 import akka.http.scaladsl.model.Uri
@@ -13,17 +13,18 @@ import akka.stream.scaladsl.Sink
 import scala.concurrent.Future
 import akka.stream.alpakka.s3.impl.CompleteMultipartUploadResult
 
-case class MultipartUploadResult(location: Uri, bucket: String, key: String, etag: String)
+final case class MultipartUploadResult(location: Uri, bucket: String, key: String, etag: String)
 
 object MultipartUploadResult {
-  def apply(r: CompleteMultipartUploadResult) = new MultipartUploadResult(r.location, r.bucket, r.key, r.etag)
+  def apply(r: CompleteMultipartUploadResult): MultipartUploadResult =
+    new MultipartUploadResult(r.location, r.bucket, r.key, r.etag)
 }
 
 object S3Client {
   val MIN_CHUNK_SIZE = 5242880
 }
 
-class S3Client(credentials: AWSCredentials, region: String)(implicit system: ActorSystem, mat: Materializer) {
+final class S3Client(credentials: AWSCredentials, region: String)(implicit system: ActorSystem, mat: Materializer) {
   import S3Client._
   private val impl = new S3Stream(credentials, region)
