--- s3/src/main/scala/akka/stream/alpakka/s3/signing/Utils.scala
@@ -1,15 +1,0 @@
-package akka.stream.alpakka.s3.signing
-
-import javax.xml.bind.DatatypeConverter
-
-import akka.util.ByteString
-
-object Utils {
-  def encodeHex(bytes: Array[Byte]): String = {
-    DatatypeConverter.printHexBinary(bytes).toLowerCase
-  }
-
-  def encodeHex(bytes: ByteString): String = {
-    encodeHex(bytes.toArray)
-  }
-}
