+++ s3/src/test/resources/application.conf
@@ -1,3 +1,3 @@
 akka {
-  loglevel = "DEBUG"
+  loggers = ["akka.testkit.TestEventListener"]
 }
